requirements.txt
# requirements.txt
ttkbootstrap
pandas
matplotlib
python-dateutil
Pillow